+++ plugins/newarc/src/plugins/7z.all/detect/isodetect.cpp
@@ -30,7 +30,7 @@ int IsIsoHeader(const unsigned char *Data,int DataSize)
 	//for (const unsigned char *CurData=Data; CurData<MaxData; CurData++)
 	//{
 		if ( IsValidHeader (CurData) )
-			return CurData-Data;
+			return (int)(CurData-Data);
 	//}
 	return -1;
 }
